pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
    "setuptools_scm[toml]>=7"
]
build-backend = "setuptools.build_meta"

[project]
name = "legend_pygeom_l1000"
description = "Geometry model of the LEGEND-1000 experimental setup"
authors = [
    { name = "Moritz Neuberger", email = "moritz.neuberger@tum.de" },
]
maintainers = [
    { name = "The LEGEND Collaboration" },
]
readme = "README.md"
license.file = "LICENSE"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: MacOS",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.10"
dependencies = [
    "legend-pygeom-hpges >=0.7.0,<0.9",
    "legend-pygeom-optics >=0.13,<0.15",
    "legend-pygeom-tools ==0.0.24",
    "numpy",
    "pint",
    "pyg4ometry",
    "dbetto",
    "pylegendmeta >=1.3.1",
    "pyyaml",
    "scipy",
]
dynamic = [
    "version",
]

[project.urls]
Homepage = "https://github.com/legend-exp/legend-pygeom-l1000"
"Bug Tracker" = "https://github.com/legend-exp/legend-pygeom-l1000/issues"
Discussions = "https://github.com/legend-exp/legend-pygeom-l1000/discussions"
Changelog = "https://github.com/legend-exp/legend-pygeom-l1000/releases"

[tool.setuptools_scm]
write_to = "src/pygeoml1000/_version.py"

[project.optional-dependencies]
all = [
    "legend-pygeom-l1000[docs,test]",
]
docs = [
    "furo",
    "myst-parser",
    "sphinx",
    "sphinx-copybutton",
    "sphinx-subfigure",
]
test = [
    "pre-commit",
    "pylegendtestdata",
    "pytest>=6.0",
    "pytest-cov",
]

[project.scripts]
legend-pygeom-l1000 = "pygeoml1000.cli:dump_gdml_cli"

[tool.setuptools]
include-package-data = true
zip-safe = false
license-files = [
    "LICENSE",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = [
    "src",
]
namespaces = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = "error"
log_cli_level = "INFO"
testpaths = "tests"

[tool.codespell]
ignore-words-list = "manuel,tre,crate"

[tool.ruff]
src = ["src"]
line-length = 110

[tool.ruff.lint]
extend-select = [
  "B",        # flake8-bugbear
  "I",        # isort
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "ICN",      # flake8-import-conventions
  "G",        # flake8-logging-format
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
  "EXE",      # flake8-executable
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
]
ignore = [
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
  "PT011",    # too picky pytest.raises() complaint
  "PLC2401",  # we like greek letters
  "PLC2403",  # we _still_ like greek letters
  "RUF003",   # we like greek letters ("σ" _of course_ totally looks like an "o"...) and math symbols like "×"
]
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]

[tool.pylint]
py-version = "3.9"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
  "too-many-nested-blocks"
]
